+++ stores/store.go
@@ -3,9 +3,9 @@ package stores
 import "github.com/roboncode/go-urlshortener/models"
 
 type Store interface {
-	IncCount() int64
+	IncCount() int
 	Create(code string, url string) (*models.Link, error)
 	Read(code string) (*models.Link, error)
-	List(limit int64, skip int64) []models.Link
-	Delete(code string) int64
+	List(limit int, skip int) []models.Link
+	Delete(code string) int
 }
